requirements.txt
python-telegram-bot==21.6
ccxt==4.4.27
pandas==1.5.3
numpy==1.24.4
matplotlib==3.7.3
mplfinance==0.12.10b0
python-dotenv==1.0.1
